Bug fix when adding adding variance planes form other ADs to current variance extension in current AD
git-svn-id: http://chara.hi.gemini.edu/svn/DRSoftware/gemini_python/trunk@4082 e04282df-fb77-4b70-bf7a-c1223586725c

+++ astrodata_Gemini/RECIPES_Gemini/primitives/primitives_standardize.py
@@ -721,8 +721,8 @@ class StandardizePrimitives(GENERALPrimitives):
                                  "with the current variance extension "
                                  "%s[%s,%d]" % (adinput.filename, VAR, extver))
                     adinput[VAR, extver].data = np.add(
-                      adinput[VAR, extver].data, var_array_final,
-                      dtype=var_dtype)
+                      adinput[VAR, extver].data,
+                      var_array_final).astype(var_dtype)
             else:
                 # Create the variance AstroData object
                 var = AstroData(data=var_array_final)
